#LMS-183 Changed optlock to Long object instead of basic type int
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/DWOProject/DWO-LMS/DWO-commons-lib/trunk@19466 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/main/java/fi/dwo/commons/persistence/entities/PersistentAppletConfig.java
@@ -69,7 +69,7 @@ public class PersistentAppletConfig implements Serializable {
 	 */
 	@Column(name = "optlock")
 	@Version
-	private int optlock;
+	private Long optlock;
 	@Column(name = "lastChangeTimeStamp")
 	private long lastChangeTimeStamp;
 	@NotNull
